Refactor Resolve TypeRefs for Policy
